requirements.txt
Django==4.0.2
djangorestframework==3.13.1
drf_yasg==1.20.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
django-environ==0.8.1
djangorestframework_simplejwt==5.0.0
mysqlclient==2.1.0